+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# shepherd 0.1.3 (18-09-2011)
+
++ Now we can `show` some sheeps!
++ `init` command now saves the inited projects!
++ Provide setup utility.
+
 # shepherd 0.1.2 (18-07-2011)
 
 + Make the `Counter` class